test: complete current-plugin-metadata-snapshot mock factories
Explicit vi.mock factories for src/plugins/current-plugin-metadata-snapshot.ts
exported only getCurrentPluginMetadataSnapshot; under isolate:false shard
composition the incomplete mocks leaked into sibling files (CI incident:
model-resolution-consistency.test.ts lost withPluginMetadataSnapshotScope,
stopgapped in #129526). Convert all 27 incomplete factories to the
importOriginal-spread form, overriding only the stubbed binding, and remove
the now-unneeded self-defense identity mock.

model.static-catalog.test.ts sat at the 1000-counted-line max-lines cap, so
its canonicalizeManifestModelCatalogProviderAlias block moved to a new
model.static-catalog.provider-alias.test.ts sibling (picked up by the
agents-embedded lane glob) instead of adding a suppression.
