+++ src/agents/model-auth-env.provider-aliases.test.ts
@@ -42,7 +42,8 @@ const setupRegistryMocks = vi.hoisted(() => ({
   resolvePluginSetupProviderCore: vi.fn(() => undefined),
 }));
 
-vi.mock("../plugins/current-plugin-metadata-snapshot.js", () => ({
+vi.mock("../plugins/current-plugin-metadata-snapshot.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../plugins/current-plugin-metadata-snapshot.js")>()),
   getCurrentPluginMetadataSnapshot: pluginMetadataMocks.getCurrentPluginMetadataSnapshot,
 }));
