+++ src/agents/provider-attribution.catalog-endpoints.test.ts
@@ -3,10 +3,11 @@ import { describe, expect, it, vi } from "vitest";
 
 // Simulates a built dist tree: externalized provider metadata comes from the
 // catalog, while one installed manifest proves first-match precedence.
-vi.mock("../plugins/current-plugin-metadata-snapshot.js", async () => {
+vi.mock("../plugins/current-plugin-metadata-snapshot.js", async (importOriginal) => {
   const { buildPluginMetadataProviderFacts } =
     await import("../plugins/plugin-metadata-provider-facts.js");
   return {
+    ...(await importOriginal<typeof import("../plugins/current-plugin-metadata-snapshot.js")>()),
     getCurrentPluginMetadataSnapshot: () => ({
       owners: buildPluginMetadataProviderFacts([
         {
