+++ src/commands/doctor/shared/stale-auth-order.test.ts
@@ -54,7 +54,10 @@ const pluginMetadataMocks = vi.hoisted(() => {
   };
 });
 
-vi.mock("../../../plugins/current-plugin-metadata-snapshot.js", () => ({
+vi.mock("../../../plugins/current-plugin-metadata-snapshot.js", async (importOriginal) => ({
+  ...(await importOriginal<
+    typeof import("../../../plugins/current-plugin-metadata-snapshot.js")
+  >()),
   getCurrentPluginMetadataSnapshot: pluginMetadataMocks.getCurrentPluginMetadataSnapshot,
 }));
