+++ src/media-understanding/defaults.test.ts
@@ -97,7 +97,8 @@ vi.mock("../plugins/manifest-contract-eligibility.js", () => ({
   }),
 }));
 
-vi.mock("../plugins/current-plugin-metadata-snapshot.js", () => ({
+vi.mock("../plugins/current-plugin-metadata-snapshot.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../plugins/current-plugin-metadata-snapshot.js")>()),
   getCurrentPluginMetadataSnapshot: () => ({
     plugins: mediaMetadataPlugins,
   }),
