+++ src/secrets/provider-env-vars.dynamic.test.ts
@@ -128,7 +128,8 @@ function useRegistrySetupPlugin(id: string, origin: string, provider: MockSetupP
   useRegistryPlugins(setupPlugin(id, origin, provider));
 }
 
-vi.mock("../plugins/current-plugin-metadata-snapshot.js", () => ({
+vi.mock("../plugins/current-plugin-metadata-snapshot.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../plugins/current-plugin-metadata-snapshot.js")>()),
   getCurrentPluginMetadataSnapshot: pluginRegistryMocks.getCurrentPluginMetadataSnapshot,
 }));
